litellm-example/requirements.txt
litellm
python-dotenv
